fix: add codeblock commands to highlighted extension
